Fix syntax for first note in a chord
Summary:
This affects the syntax highlighting of a chord like `<c e g>`---with no
initial whitespace, as opposed to `< c e g >`. Before this patch, only
the `e` and `g` notes were matched as `lilyNote`s (and thus
highlighted). This is because the Scheme syntax settings mark `<` as a
keyword character, which interferes with the matching. This patch
prevents that by saving and restoring the value of the 'iskeyword'
variable before and after we include the Scheme syntax file.

Test Plan:
In a file `test.ly`, write

    <c e g>
    < c e g >

and make sure that all letters are marked as identifiers. (Before this
patch, only the last five letters were; the initial `c` was not.)

Also, for good measure, note that a Lilypond file containing

    #(define pi 3.2)

is still highlighted correctly.