Remove obtrusive `showmatch` from ftplugin
Summary:
This setting has nothing to do with Lilypond and seems to be based on
the author's personal preferences. For me, this setting is definitely
not desired, as it introduces a perceptual lag: when typing the sequence
"open-brace newline close-brace," the close-brace is only dedented after
a time of `matchtime` deciseconds.

If users want this to be set, they should set it in an ftplugin autocmd.
It should not be part of the Lilypond bundle.