Greatly improve syntax highlighting performance
Summary:
The part of syntax highlighting that takes by far the longest is the
matching of `lilyNote`s. Not only is the corresponding regular
expression long, it is also complex, with multiple opportunities for
backtracking. (For comparison, the regular expression for `lilyKeyword`
and `lilyReservedWord` are also long, but very simple, and do not slow
down highlighting by much.)

This patch simply strips down the regular expression by only including
the core family of note types unless an option is set. That is, notes
like `csharpsharp` (for `cisis`) will no longer be recognized by
default.

After this patch, scrolling through a buffer has no lag on my machine,
even with `cursorline` enabled.

For the record, I also tried exploding `lilyKeyword` and
`lilyReservedWord` into `syn-keyword`s, but (a) this wasn't faithful to
the original semantics because underscores and backslashes can interfere
(so `c_\markup { legato }` failed to mark `\markup` as a keyword), and
(b) it didn't seem to have any significant performance improvement,
anyway.